lindera::token_filter

Trait TokenFilter

pub trait TokenFilter:
    'static
    + Send
    + Sync
    + TokenFilterClone {
    // Required methods
    fn name(&self) -> &'static str;
    fn apply(&self, tokens: &mut Vec<Token<'_>>) -> LinderaResult<()>;
}

A trait for token filters that can be applied to a vector of tokens.

This trait requires the implementor to be 'static, Send, Sync, and implement the TokenFilterClone trait. It provides methods to get the name of the filter and to apply the filter to a mutable vector of tokens.

Required Methods

name: Returns the name of the token filter as a static string slice.

apply: Applies the token filter to a mutable vector of tokens, returning a LinderaResult<()>.

Required Methods

fn name(&self) -> &'static str

fn apply(&self, tokens: &mut Vec<Token<'_>>) -> LinderaResult<()>